refactor(pipeline/scrape): extract BfsFieldSearch + TxnSignature
Phase 5 commit 3/9 - move the BFS-based field-discovery engine and
the transaction-signature classifier out of ScrapeAutoMapper.ts:

- BfsFieldSearch/BfsFieldSearch.ts (138 eff): ISearchItem,
  isSearchableObject, tryMatchWk, matchField, matchFieldInRecord,
  enqueueChildren, bfsOneLevel, bfsAccumulate, flattenObjectTree,
  findFieldValue, collectFieldValuesDeep + ...FromRecord,
  findAllFieldValues. Imports ScraperErrorTypes, IFieldMatch,
  Procedure helpers, MAX_SEARCH_DEPTH.

- BfsFieldSearch/TxnSignature.ts (23 eff): MIN_TXN_SCORE,
  TXN_SIGNATURE_FIELDS, scoreTxnSignature, castSearchable. Sub-split
  out so BfsFieldSearch.ts stays under the per-cluster max-lines:150
  eff cap — both files are independently under the cap.

ScrapeAutoMapper.ts (617 eff, was 752) now imports the surface
from the two new modules and drops the now-unused ScraperErrorTypes,
IFieldMatch, Procedure, fail/isOk/succeed, MAX_SEARCH_DEPTH imports.

Public surface (findAllFieldValues, findFieldValue, matchField)
is unchanged — still re-exported via the bottom named-export block.

Master plan: pipeline-decoupling-master-2026-05-28 / phase-5
Cap: max-lines 150 eff + max-lines-per-function 20 eff
File eff LoC: BfsFieldSearch.ts=138, TxnSignature.ts=23 (cap 150)
